test(vault): fold the new tests into the existing test files
Addresses review feedback: interceptor_extra_test.go and
vault_extra_test.go duplicated files that already existed. The test
bodies move unchanged into interceptor_test.go and vault_test.go.

Refs #1758
